🔨 [projects] Move function `storeproject` to `store` module
