+++ src/cutty/projects/generate.py
@@ -2,11 +2,9 @@
 import pathlib
 from collections.abc import Sequence
 
-from cutty.filestorage.adapters.cookiecutter import createcookiecutterstorage
 from cutty.filestorage.adapters.disk import FileExistsPolicy
-from cutty.filesystems.domain.purepath import PurePath
 from cutty.projects.generator import ProjectGenerator
-from cutty.projects.project import Project
+from cutty.projects.store import storeproject
 from cutty.projects.template import Template
 from cutty.templates.adapters.cookiecutter.binders import bindcookiecuttervariables
 from cutty.templates.domain.bindings import Binding
@@ -43,26 +41,3 @@ def generate(
         outputdirisproject,
         fileexists,
     )
-
-
-def storeproject(
-    project: Project,
-    outputdir: pathlib.Path,
-    outputdirisproject: bool,
-    fileexists: FileExistsPolicy,
-) -> pathlib.Path:
-    """Store a project in the output directory."""
-    projectdir = outputdir if outputdirisproject else outputdir / project.name
-    storage = createcookiecutterstorage(
-        outputdir, projectdir, fileexists, project.hooks
-    )
-
-    with storage:
-        for projectfile in project.files:
-            if outputdirisproject:
-                path = PurePath(*projectfile.path.parts[1:])
-                projectfile = projectfile.withpath(path)
-
-            storage.add(projectfile)
-
-    return projectdir
